moved resource IDs from common to common::res_id
