+++ src/common/wrapper.rs
@@ -14,7 +14,14 @@ use xrbk::{
 	X11Size,
 };
 
-use crate::{Atom, Button, Colormap, Keycode, Pixmap, Timestamp, Window, WindowClass};
+use crate::{
+	atom::Atom,
+	res_id::{Colormap, Pixmap, Window},
+	Button,
+	Keycode,
+	Timestamp,
+	WindowClass,
+};
 
 macro_rules! impl_constant_x11_size { // {{{
 	($type:ty {
